RDCC-5936:Changes made to add initial configuration and adding LD changes
