+++ src/main/java/uk/gov/hmcts/reform/elinks/util/RefDataConstants.java
@@ -9,5 +9,12 @@ public class RefDataConstants {
     public static final String AUTHORIZATION = "Authorization";
     public static final String CONTENT_TYPE = "Content-Type";
 
+    public static final String BAD_REQUEST = "Bad Request";
+    public static final String FORBIDDEN_ERROR = "Forbidden Error: Access denied for invalid permissions";
+    public static final String UNAUTHORIZED_ERROR =
+            "Unauthorized Error : The requested resource is restricted and requires authentication";
+    public static final String NO_DATA_FOUND = "The eLinks data could not be found";
+    public static final String INTERNAL_SERVER_ERROR = "Internal Server Error";
+
 
 }
